+++ sample/lib/widgets/showcase/sample_user_avatar.showcas.dart
@@ -1,0 +1,34 @@
+import 'package:blip_ds/blip_ds.dart';
+import 'package:flutter/material.dart';
+
+class SampleUserAvatarShowcase extends StatelessWidget {
+  const SampleUserAvatarShowcase({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: [
+        DSUserAvatar(
+          text: "Adriano Chamon Tavares",
+        ),
+        const SizedBox(
+          height: 15,
+        ),
+        DSUserAvatar(
+          text: "Adriano Chamon Tavares",
+          textStyle: const DSHeadlineLargeTextStyle(),
+        ),
+        const SizedBox(
+          height: 15,
+        ),
+        DSUserAvatar(
+          text: "Adriano Chamon Tavares",
+          textStyle: const DSHeadlineExtraLargeTextStyle(),
+        ),
+        const SizedBox(
+          height: 15,
+        ),
+      ],
+    );
+  }
+}